feat: use ZipFileEntry instead of ZipStreamFileMetadata
